refactor: rename start time entry dto
